Serializing object to JSON allows us to parse them
Before they were just put in as a String and we
had no other chance then `eval` those to the
corresponding object. Quite spooky and dangerous!

None of the Redis gems that we are using right now
has the ability to use other data types then
strings. So this approach is the one without the
pain and new adventures we might get in exchange
for more crowded type support at Redis side.